+++ Nbm/src/test/resources/projForTest/src/intentions/invertIfCondition.kt
@@ -1,0 +1,9 @@
+package intentions
+
+fun test(x: Int): String {
+    if (x > 0) {
+        return "positive"
+    } else {
+        return "non-positive"
+    }
+}